Initial release of LlamaCards using WebGPU and Web-Llm
Simply open this .html file in your browser to try out LlamaCards.  

Please keep in mind this is just a demo example with a hardcoded Gemma2:2b model.
Model cards are not processed in parallel or concurrently like with the Ollama backend version.  

Base functionality is here but layouts/panning and zooming are broken at the moment.